엔티티 정리

+++ backend/src/main/java/com/team7/ConcerTUNE/entity/ArtistGenreId.java
@@ -18,5 +18,5 @@ public class ArtistGenreId implements Serializable {
     private Long artistId;
 
     @Column(name = "genre_id")
-    private Integer genreId;
+    private Long genreId;
 }
